# _oas/openapi.yaml
openapi: 3.0.3
info:
  title: Go Faster Status API
  description: Statistical insights for thr Go Faster system
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://github.com/go-faster/bot/blob/main/LICENSE
servers:
  - url: https://api.go-faster.org/
    description: Production
  - url: https://localhost:8081
    description: Local
paths:
  /status:
    get:
      operationId: "status"
      description: "get status"
      responses:
        200:
          description: Go Faster Status
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Status"
        default:
          $ref: "#/components/responses/Error"
  /badge/telegram/online:
    parameters:
      - required: true
        in: query
        name: groups
        schema:
          type: array
          items:
            type: string
    get:
      operationId: "getTelegramOnlineBadge"
      responses:
        200:
          $ref: "#/components/responses/SVG"
        default:
          $ref: "#/components/responses/Error"
  /badge/telegram/{group_name}:
    parameters:
      - required: true
        in: path
        name: group_name
        schema:
          type: string
    get:
      operationId: "getTelegramBadge"
      description: "get svg badge for telegram group"
      parameters:
        - in: query
          name: title
          required: false
          schema:
            type: string
      responses:
        200:
          $ref: "#/components/responses/SVG"
        default:
          $ref: "#/components/responses/Error"
  /github/status:
    post:
      description: https://www.githubstatus.com/ webhook
      operationId: "githubStatus"
      parameters:
        - name: secret
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: object
              additionalProperties: true
      responses:
        200:
          description: "Github status accepted"
        default:
          $ref: "#/components/responses/Error"

components:
  responses:
    Error:
      description: "Error while processing request"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/Error"
    SVG:
      description: Badge
      headers:
        Cache-Control:
          description: "The Cache-Control header provides important info on how long a response may be considered fresh"
          schema:
            type: string
          example: "max-age=3600, public, must-revalidate"
        ETag:
          description: "The ETag header provides a unique identifier for the response"
          schema:
            type: string
          example: "W/\"123456789\""
      content:
        "image/svg+xml":
          schema:
            type: string
            format: binary
  schemas:
    Error:
      type: object
      description: "error description"
      properties:
        message:
          type: string
      required: [ message ]
    User:
      type: object
      required:
        - id
        - nickname
        - full_name
        - avatar_url
        - html_url
      properties:
        id:
          type: integer
          format: int64
          description: User ID
        nickname:
          type: string
          description: User name
        full_name:
          type: string
          description: User full name
        avatar_url:
          type: string
          format: uri
          description: User avatar URL
        html_url:
          type: string
          format: uri
          description: User profile URL
    UserStat:
      type: object
      description: User statistics
      required:
        - user
        - total_commits
      properties:
        user:
          $ref: "#/components/schemas/User"
        total_commits:
          type: integer
          description: Total commits
    Statistics:
      type: object
      description: Statistics
      required:
        - total_commits
        - top_users
      properties:
        total_commits:
          type: integer
          description: Total commits
        top_users:
          type: array
          description: Top users
          items:
            $ref: "#/components/schemas/UserStat"
    Status:
      type: object
      description: Status
      required: [ stat, message ]
      properties:
        message:
          type: string
          description: Message
        stat:
          $ref: "#/components/schemas/Statistics"